- reorganize database configurators and providers
