+++ .nox/design/sample.solution.nox.yaml
@@ -412,19 +412,19 @@ infrastructure:
       name: SampleCurrencyDb
 
       # Sql Server
-      #serverUri: localhost
-      #provider: sqlServer
-      #port: 1433
-      #user: sa
-      #password: Developer*123
-      #options: Trusted_Connection=no;connection timeout=120;TrustServerCertificate=True;
+      serverUri: localhost
+      provider: sqlServer
+      port: 1433
+      user: sa
+      password: Developer*123
+      options: Trusted_Connection=no;connection timeout=120;TrustServerCertificate=True;
 
       # Postgres
-      serverUri: localhost
-      provider: postgres
-      port: 54320
-      user: sa
-      password: Developer*123      
+      #serverUri: localhost
+      #provider: postgres
+      #port: 54320
+      #user: sa
+      #password: Developer*123      
       
       ### MySql  
       #provider: mySql
