+++ src/Nox.DatabaseProvider.Postgres/Nox.DatabaseProvider.Postgres.csproj
@@ -7,11 +7,11 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Nox.DatabaseProvider\Nox.DatabaseProvider.csproj" />
+      <ProjectReference Include="..\Nox.Types.EntityFramework\Nox.Types.EntityFramework.csproj" />
     </ItemGroup>
 
     <ItemGroup>
       <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="7.0.4" />
     </ItemGroup>
-    
+
 </Project>
