+++ src/Nox.DatabaseProvider.Sqlite/Nox.DatabaseProvider.Sqlite.csproj
@@ -10,8 +10,4 @@
       <ProjectReference Include="..\Nox.Types.EntityFramework\Nox.Types.EntityFramework.csproj" />
     </ItemGroup>
 
-    <ItemGroup>
-      <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.8" />
-    </ItemGroup>
-
 </Project>
