--- src/Nox.DatabaseProvider/Nox.DatabaseProvider.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.8" />
-      <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="7.0.8" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Nox.Abstractions\Nox.Abstractions.csproj" />
-      <ProjectReference Include="..\Nox.Solution\Nox.Solution.csproj" />
-    </ItemGroup>
-
-</Project>
