+++ src/Nox.Lib/Extensions/ServiceCollectionExtension.cs
@@ -1,6 +1,5 @@
 using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
-using Nox.DatabaseProvider;
 using Nox.Solution;
 
 namespace Nox;
