+++ src/Nox.Types.EntityFramework.SqlServer/Nox.Types.EntityFramework.SqlServer.csproj
@@ -20,7 +20,7 @@
 <!--	</ItemGroup>-->
 
 	<ItemGroup>
-		<ProjectReference Include="..\Nox.Types.EntityFramework.vNext\Nox.Types.EntityFramework.vNext.csproj" />
+	  <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.8" />
 	</ItemGroup>
 
 </Project>
