+++ src/Nox.Types.EntityFramework.Sqlite/Nox.Types.EntityFramework.Sqlite.csproj
@@ -19,8 +19,4 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="7.0.8" />
   </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Nox.Types.EntityFramework.vNext\Nox.Types.EntityFramework.vNext.csproj" />
-  </ItemGroup>
-
 </Project>
