+++ src/Nox.Types.EntityFramework.vNext/Exceptions/NoxEntityFrameworkException.cs
@@ -1,4 +1,4 @@
-﻿namespace Nox.Types.EntityFramework.vNext
+﻿namespace Nox.Types.EntityFramework.vNext.Exceptions
 {
     public class NoxEntityFrameworkException : Exception
     {
