+++ src/Nox.Types.EntityFramework.vNext/Nox.Types.EntityFramework.vNext.csproj
@@ -24,4 +24,8 @@
 	  <ProjectReference Include="..\Nox.Types.EntityFramework\Nox.Types.EntityFramework.csproj" />
 	</ItemGroup>
 
+	<ItemGroup>
+	  <Folder Include="Types\" />
+	</ItemGroup>
+
 </Project>
