+++ src/Nox.Types.EntityFramework/Abstractions/INoxDatabaseConfigurator.cs
@@ -1,0 +1,9 @@
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using Nox.Solution;
+
+namespace Nox.Types.EntityFramework.Abstractions;
+
+public interface INoxDatabaseConfigurator
+{
+    void ConfigureEntity(EntityTypeBuilder builder, Entity entity);
+}
