+++ src/Nox.Types.EntityFramework/Configurators/NumberDatabaseConfigurator.cs
@@ -1,8 +1,9 @@
 ﻿using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Nox.Solution;
+using Nox.Types.EntityFramework.Abstractions;
 using Nox.Types.EntityFramework.Types;
 
-namespace Nox.Types.EntityFramework.vNext.Types;
+namespace Nox.Types.EntityFramework.Configurators;
 
 public class NumberDatabaseConfigurator : INoxTypeDatabaseConfigurator
 {
