+++ src/Nox.Types.EntityFramework/Configurators/TextDatabaseConfigurator.cs
@@ -1,9 +1,10 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Nox.Solution;
+using Nox.Types.EntityFramework.Abstractions;
 using Nox.Types.EntityFramework.Types;
 
-namespace Nox.Types.EntityFramework.vNext.Types;
+namespace Nox.Types.EntityFramework.Configurators;
 
 public class TextDatabaseConfigurator : INoxTypeDatabaseConfigurator
 {
